+++ config/application.php
@@ -18,7 +18,6 @@ use Opulence\Ioc\IContainer;
  * ----------------------------------------------------------
  */
 $taskDispatcher = require __DIR__ . "/tasks.php";
-$environment = require __DIR__ . "/environment.php";
 $container = require __DIR__ . "/ioc.php";
 $application = new Application($taskDispatcher, $environment);