+++ config/exceptionHandler.php
@@ -6,6 +6,8 @@ use Opulence\Http\HttpException;
  * ----------------------------------------------------------
  * Define the exception handler
  * ----------------------------------------------------------
+ *
+ * The last parameter lists any exceptions you do not want to log
  */
 return new ExceptionHandler(
     require __DIR__ . "/logging.php",