+++ configs/console/exceptionRenderer.php
@@ -1,9 +1,0 @@
-<?php
-use Opulence\Framework\Debug\Exceptions\Handlers\Console\ExceptionRenderer;
-
-/**
- * ----------------------------------------------------------
- * Set the console exception renderer
- * ----------------------------------------------------------
- */
-return new ExceptionRenderer();