+++ src/Project/Bootstrappers/Http/Sessions/SessionBootstrapper.php
@@ -97,7 +97,7 @@ class SessionBootstrapper extends BaseBootstrapper
     private function loadConfig()
     {
         if ($this->config === null) {
-            $this->config = require "{$this->paths["configs.http"]}/sessions.php";
+            $this->config = require "{$this->paths["config.http"]}/sessions.php";
         }
     }
 }