+++ src/Project/Bootstrappers/Http/Views/ViewBootstrapper.php
@@ -20,7 +20,7 @@ class ViewBootstrapper extends BaseBootstrapper
      */
     protected function getViewCache(IContainer $container)
     {
-        $cacheConfig = require_once "{$this->paths["configs.http"]}/views.php";
+        $cacheConfig = require_once "{$this->paths["config.http"]}/views.php";
 
         return new FileCache(
             $this->paths["views.compiled"],