+++ src/Project/Http/Middleware/CheckCsrfToken.php
@@ -46,7 +46,7 @@ class CheckCsrfToken extends BaseMiddleware
     private function loadConfig()
     {
         if ($this->config === null) {
-            $this->config = require "{$this->paths["configs.http"]}/sessions.php";
+            $this->config = require "{$this->paths["config.http"]}/sessions.php";
         }
     }
 }